backend/requirements.txt
# Django Framework
django>=4.2
djangorestframework>=3.14
django-cors-headers>=4.3

# Data Processing
pandas>=2.0
numpy>=1.24

# Machine Learning
scikit-learn>=1.3
joblib>=1.3

# For handling imbalanced datasets (fraud detection)
imbalanced-learn>=0.11
